Update rscache dependencies and test packaging
