+++ pkg/rscache/internal/integration_test/go.mod
@@ -1,11 +1,13 @@
-module github.com/rstudio/platform-lib/pkg/rscache
+module github.com/rstudio/platform-lib/pkg/rscache/internal/integration_test
 
 go 1.17
 
 require (
 	github.com/dgraph-io/ristretto v0.1.0
 	github.com/fortytw2/leaktest v1.3.0
+	github.com/rstudio/platform-lib/pkg/rscache v0.1.0
 	github.com/rstudio/platform-lib/pkg/rsstorage v0.3.0
+	github.com/rstudio/platform-lib/pkg/rsstorage/servers/file v0.3.1
 	gopkg.in/check.v1 v1.0.0-20201130134442-10cb98267c6c
 )
 
@@ -16,8 +18,10 @@ require (
 	github.com/golang/glog v0.0.0-20160126235308-23def4e6c14b // indirect
 	github.com/kr/pretty v0.2.1 // indirect
 	github.com/kr/text v0.2.0 // indirect
+	github.com/minio/minio v0.0.0-20210323145707-da70e6ddf63c // indirect
+	github.com/montanaflynn/stats v0.5.0 // indirect
+	github.com/ncw/directio v1.0.5 // indirect
 	github.com/pkg/errors v0.9.1 // indirect
-	github.com/stretchr/testify v1.6.1 // indirect
+	github.com/stretchr/testify v1.7.0 // indirect
 	golang.org/x/sys v0.0.0-20210615035016-665e8c7367d1 // indirect
-	gopkg.in/yaml.v3 v3.0.0-20200615113413-eeeca48fe776 // indirect
 )
